POC script organiser.
Adds a simple independent script_organiser attribute to the ScriptRunner
class, which reads the attribute (dict) CALLBACK_ORDER from scripts
should it exist, and organises script running order by priority value
assigned to each callback, instead of the longstanding name of
extension's folder. Priority is optional and currently defaults to 50 as
a placeholder, and I've only applied it to process callback's alwayson
scripts, but it seems to work correctly.
This allows fine grained control for extensions to choose when to run
each callback relative to other extensions, which is crucial for several
extensions' correct operation (controlnet, animatediff, regional
prompter, adetailer, ultimate upscale, wildcards, segment anything to
name a few). Currently this order is global and indirectly user
controlled (naming scheme), which is really bad for devs.