Thumbnail Resolution and Generation Refactor
Recent changes to how Thumbnails were generated aimed to reduce file-system access, and in turn created a logarythmic computational load.

This refactor moves the vast majority of Thumbnail Resolution and Generation into a single thread.
I can not give a "%" improvement, as I do not have the patience to wait for a load with base-dev at this time.
That said, I have a very large collection (well in excess of 500K files when including images and info, all over a network) and startup with this commit is roughly 230 seconds.
I *will not* get a base metwric to compare without, but suffice it to say I aborted the load after over 10 minutes.  IIRC, my initial loads we in excess of 20 minutes prior to working on this commit.

The added labor of resolving the thumbnail at request time (refactor of the `fetch_file` api call) does not incur a measurable impact on page load time.

This commit also includes a refactor in how the HTML for cards is generated.  This reduced page-size considerably in my testing, but may not be as profound in most cases (my library consists of over 50k models).